core(scheduler) + test cases: implement RealScheduler and boundary tests
- Add RealScheduler with Start loop, drift correction, and context cancel
- Add Ticks() read-only channel and durationFor() helper
- Unify nextboundary to use durationFor() for all frequencies
- Fix Day boundary: 24+time.Hour → 24*time.Hour operator bug
- Add FakeTimeSource for hermetic time control in tests
- Add TestNextBoundary tests for Second, Minute, Hour, Day
- Add TestStart: tick emission and no-consumer deadlock guard

+++ cmd/producer/main.go
@@ -9,7 +9,7 @@ import (
 
 type ProducerStruct struct {
 	version int
-	uuid    int
+	uuid    uint8
 }
 
 type Event struct {
@@ -17,7 +17,7 @@ type Event struct {
 	timestamp       time.Time
 	FrequencyType   string
 	monotonicSeqNum int
-	//Deterministic Signal Seed
+
 	Producer ProducerStruct
 	Version  string
 }
@@ -29,5 +29,4 @@ func main() {
 	for range ticker.C {
 		fmt.Println(ts.Now())
 	}
-
 }
